Predetermine block flags and add get_modes function
Perdetermine flags for each block (whether it is opaque or non-opaque,
whether it uses a maximum of two colors etc), to possibly limit the
number of compression modes that has to be probed.

Add a get_modes function for each compression format that return a
list of modes based on the given block flags.

Remove the compressed block get_mode function from the compression
info table.

Signed-off-by: Harm Hanemaaijer <fgenfb@yahoo.com>